requirements.txt
flask
huggingface_hub
python-dotenv
markdown
markupsafe
